Fix normalization transformers not using correct/updated recipe references
If multiple transform functions are used, the first may potentially copy and create new recipe references that are traversed later. Future transform functions will then be operating on stale recipe references. This fixes that by creating the notion of a `TransformBinder` which is given the current "root recipe" in the traversal loop, fixing this issue.