Redesign `aheadIs()` splicing to "hoist" the recipe outside behind-like recipes
This fixes an issue where cases like applying `behindIs()` on `asdf(?=...)` resulted in `(?<=asdf...)`, which led to unexpected results. If this "asdf rule" was checked immediately behind a "..." in "asdf...", it would fail, because `(?<=asdf...)` does not match "asdf". Instead, this should be `(?<=asdf)(?=...)`, so the example can succeed (as expected).

This required quite a big lift and is definitely a breaking change. Current tests were modified (and added upon slightly) to compensate for this new behavior, but there are still quirks with it (like many redundant captures).