WIP: Fix many `aheadIs()` errors by making it a best-effort attempt (rather than a strict rule)
This commit makes it so `aheadIs()` recipes that aren't on the end are just removed completely, instead of resulting in an error. This really was just an experiment to see how it would work with common uses of `phrase()`s. (Spoiler alert: it kind of seems to defeat their purpose, since you now need to re-add the elements that were taken out of the `phrase()`...)